feat: install cloudinary and create basic photography page
