+++ package.json
@@ -27,6 +27,7 @@
     "@react-three/drei": "^9.122.0",
     "@react-three/fiber": "^8.18.0",
     "astro": "^5.5.5",
+    "astro-cloudinary": "^1.3.2",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "gsap": "^3.12.7",
